js/xpconnect/tests/chrome/test_bug1281071.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1281071
-->
<window title="Mozilla Bug 1281071"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
    SimpleTest.waitForExplicitFinish();
    function go() {
      var w = $('ifr').contentWindow.wrappedJSObject;
      is(w.tryLocationGet(), "Permission to call 'get location' denied.",
         "Trying to get our location object should throw");
      SimpleTest.finish();
    }
  ]]></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
    <iframe type="content"
            src="http://mochi.test:8888/chrome/js/xpconnect/tests/chrome/file_bug1281071.html"
            onload="go()"
            id="ifr">
    </iframe>
  </body>

</window>
